+++ src/ui/svelte/calendar-popup/index.ts
@@ -2,6 +2,7 @@ import { mount } from "svelte";
 import App from "./app.svelte";
 import { loadTranslations } from "./i18n/index.ts";
 import { Widget } from "@seelen-ui/lib";
+import "moment/min/locales";
 
 import "@shared/styles/reset.css";
 import "@shared/styles/colors.css";