+++ src/ui/svelte/keyboard-selector/index.ts
@@ -1,5 +1,6 @@
 import { mount } from "svelte";
 import App from "./app.svelte";
+import { loadTranslations } from "./i18n/index.ts";
 import { Widget } from "@seelen-ui/lib";
 
 import "@shared/styles/reset.css";
@@ -11,6 +12,8 @@ const widget = Widget.getCurrent();
 await widget.init({
   autoSizeByContent: root,
 });
+
+await loadTranslations();
 
 mount(App, {
   target: root,