+++ src/ui/svelte/keyboard-selector/state.svelte.ts
@@ -1,5 +1,15 @@
 import { invoke, SeelenCommand, SeelenEvent, subscribe } from "@seelen-ui/lib";
 import { lazyRune } from "libs/ui/svelte/utils";
+import { locale } from "./i18n/index.ts";
+
+let settings = lazyRune(() => invoke(SeelenCommand.StateGetSettings, { path: null }));
+subscribe(SeelenEvent.StateSettingsChanged, settings.setByPayload);
+await settings.init();
+$effect.root(() => {
+  $effect(() => {
+    locale.set(settings.value.language || "en");
+  });
+});
 
 const langs = lazyRune(() => invoke(SeelenCommand.SystemGetLanguages));
 subscribe(SeelenEvent.SystemLanguagesChanged, langs.setByPayload);